Damage, Death, Reset
I implemented an empty level that can be used to build out the first map of the forrest, named as intended as well. I also created a level manager script for all of our levels moving forward to work as a type of general purpose manager for keeping track of things like the player. Player can now die and it resets the current level.